# api/v1alpha1/openapi.yaml
openapi: 3.0.0
info:
  contact: { }
  description: KubeVirt Service Provider API
  title: Service Provider API
  version: v1alpha1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: /

paths:
  /health:
    get:
      summary: Health check
      operationId: ListHealth
      description: Health check for KubeVirt Service Provider API
      responses:
        '200':
          description: OK

  /v1/vm:
    get:
      summary: Get a virtual machine application
      operationId: GetVM
      description: Get a VM application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VM'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a VM
      operationId: createVM
      description: Create an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMRequest'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VM'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update an application
      operationId: ApplyVM
      description: Update a KubeVirt VM application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMRequest'
      responses:
        '201':
          description: Updated Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VM'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete an application
      operationId: deleteVM
      description: Delete a KubeVirt application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMRequest'
      responses:
        '204':
          description: No Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VM'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    VMRequest:
      type: object
      x-aep-resource: true
      properties:
        id:
          type: string
          description: ID of the VM
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Name of the VM
          example: "webserver"
        namespace:
          type: string
          description: Namespace of the VM
        osImage:
          type: string
          description: Base image of the OS
          example: "quay.io/containerdisks/fedora:latest"
        cpu:
          type: integer
          description: CPU for the VM
          example: 1
        ram:
          type: integer
          description: Ram for the CPU
          example: 1
    VM:
      type: object
      x-aep-resource: true
      properties:
        id:
          type: string
          description: ID of the VM
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Name of the VM
          example: "webserver"
        namespace:
          type: string
          description: Namespace of the VM
    Error:
      required:
        - error
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: integer
          description: Error code
          example: 400