# api/v1alpha1/types.gen.go
// Package v1alpha1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package v1alpha1

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code *int `json:"code,omitempty"`

	// Error Error message
	Error string `json:"error"`
}

// VM defines model for VM.
type VM struct {
	// Id ID of the VM
	Id *string `json:"id,omitempty"`

	// Name Name of the VM
	Name *string `json:"name,omitempty"`

	// Namespace Namespace of the VM
	Namespace *string `json:"namespace,omitempty"`
}

// VMRequest defines model for VMRequest.
type VMRequest struct {
	// Cpu CPU for the VM
	Cpu *int `json:"cpu,omitempty"`

	// Id ID of the VM
	Id *string `json:"id,omitempty"`

	// Name Name of the VM
	Name *string `json:"name,omitempty"`

	// Namespace Namespace of the VM
	Namespace *string `json:"namespace,omitempty"`

	// OsImage Base image of the OS
	OsImage *string `json:"osImage,omitempty"`

	// Ram Ram for the CPU
	Ram *int `json:"ram,omitempty"`
}

// DeleteVMJSONRequestBody defines body for DeleteVM for application/json ContentType.
type DeleteVMJSONRequestBody = VMRequest

// GetVMJSONRequestBody defines body for GetVM for application/json ContentType.
type GetVMJSONRequestBody = VMRequest

// CreateVMJSONRequestBody defines body for CreateVM for application/json ContentType.
type CreateVMJSONRequestBody = VMRequest

// ApplyVMJSONRequestBody defines body for ApplyVM for application/json ContentType.
type ApplyVMJSONRequestBody = VMRequest