# internal/api/server/server.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package server

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code *int `json:"code,omitempty"`

	// Error Error message
	Error string `json:"error"`
}

// VM defines model for VM.
type VM struct {
	// Id ID of the VM
	Id *string `json:"id,omitempty"`

	// Name Name of the VM
	Name *string `json:"name,omitempty"`

	// Namespace Namespace of the VM
	Namespace *string `json:"namespace,omitempty"`
}

// VMRequest defines model for VMRequest.
type VMRequest struct {
	// Cpu CPU for the VM
	Cpu *int `json:"cpu,omitempty"`

	// Id ID of the VM
	Id *string `json:"id,omitempty"`

	// Name Name of the VM
	Name *string `json:"name,omitempty"`

	// Namespace Namespace of the VM
	Namespace *string `json:"namespace,omitempty"`

	// OsImage Base image of the OS
	OsImage *string `json:"osImage,omitempty"`

	// Ram Ram for the CPU
	Ram *int `json:"ram,omitempty"`
}

// DeleteVMJSONRequestBody defines body for DeleteVM for application/json ContentType.
type DeleteVMJSONRequestBody = VMRequest

// GetVMJSONRequestBody defines body for GetVM for application/json ContentType.
type GetVMJSONRequestBody = VMRequest

// CreateVMJSONRequestBody defines body for CreateVM for application/json ContentType.
type CreateVMJSONRequestBody = VMRequest

// ApplyVMJSONRequestBody defines body for ApplyVM for application/json ContentType.
type ApplyVMJSONRequestBody = VMRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Health check
	// (GET /health)
	ListHealth(w http.ResponseWriter, r *http.Request)
	// Delete an application
	// (DELETE /v1/vm)
	DeleteVM(w http.ResponseWriter, r *http.Request)
	// Get a virtual machine application
	// (GET /v1/vm)
	GetVM(w http.ResponseWriter, r *http.Request)
	// Create a VM
	// (POST /v1/vm)
	CreateVM(w http.ResponseWriter, r *http.Request)
	// Update an application
	// (PUT /v1/vm)
	ApplyVM(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Health check
// (GET /health)
func (_ Unimplemented) ListHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Delete an application
// (DELETE /v1/vm)
func (_ Unimplemented) DeleteVM(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get a virtual machine application
// (GET /v1/vm)
func (_ Unimplemented) GetVM(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create a VM
// (POST /v1/vm)
func (_ Unimplemented) CreateVM(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Update an application
// (PUT /v1/vm)
func (_ Unimplemented) ApplyVM(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListHealth operation middleware
func (siw *ServerInterfaceWrapper) ListHealth(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// DeleteVM operation middleware
func (siw *ServerInterfaceWrapper) DeleteVM(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteVM(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetVM operation middleware
func (siw *ServerInterfaceWrapper) GetVM(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetVM(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// CreateVM operation middleware
func (siw *ServerInterfaceWrapper) CreateVM(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateVM(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// ApplyVM operation middleware
func (siw *ServerInterfaceWrapper) ApplyVM(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ApplyVM(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.ListHealth)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/v1/vm", wrapper.DeleteVM)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/vm", wrapper.GetVM)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/vm", wrapper.CreateVM)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/v1/vm", wrapper.ApplyVM)
	})

	return r
}

type ListHealthRequestObject struct {
}

type ListHealthResponseObject interface {
	VisitListHealthResponse(w http.ResponseWriter) error
}

type ListHealth200Response struct {
}

func (response ListHealth200Response) VisitListHealthResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type DeleteVMRequestObject struct {
	Body *DeleteVMJSONRequestBody
}

type DeleteVMResponseObject interface {
	VisitDeleteVMResponse(w http.ResponseWriter) error
}

type DeleteVM204JSONResponse VM

func (response DeleteVM204JSONResponse) VisitDeleteVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(204)

	return json.NewEncoder(w).Encode(response)
}

type DeleteVM400JSONResponse Error

func (response DeleteVM400JSONResponse) VisitDeleteVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type DeleteVM500JSONResponse Error

func (response DeleteVM500JSONResponse) VisitDeleteVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetVMRequestObject struct {
	Body *GetVMJSONRequestBody
}

type GetVMResponseObject interface {
	VisitGetVMResponse(w http.ResponseWriter) error
}

type GetVM200JSONResponse VM

func (response GetVM200JSONResponse) VisitGetVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetVM400JSONResponse Error

func (response GetVM400JSONResponse) VisitGetVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetVM500JSONResponse Error

func (response GetVM500JSONResponse) VisitGetVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type CreateVMRequestObject struct {
	Body *CreateVMJSONRequestBody
}

type CreateVMResponseObject interface {
	VisitCreateVMResponse(w http.ResponseWriter) error
}

type CreateVM201JSONResponse VM

func (response CreateVM201JSONResponse) VisitCreateVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateVM400JSONResponse Error

func (response CreateVM400JSONResponse) VisitCreateVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateVM500JSONResponse Error

func (response CreateVM500JSONResponse) VisitCreateVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type ApplyVMRequestObject struct {
	Body *ApplyVMJSONRequestBody
}

type ApplyVMResponseObject interface {
	VisitApplyVMResponse(w http.ResponseWriter) error
}

type ApplyVM201JSONResponse VM

func (response ApplyVM201JSONResponse) VisitApplyVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type ApplyVM400JSONResponse Error

func (response ApplyVM400JSONResponse) VisitApplyVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type ApplyVM500JSONResponse Error

func (response ApplyVM500JSONResponse) VisitApplyVMResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Health check
	// (GET /health)
	ListHealth(ctx context.Context, request ListHealthRequestObject) (ListHealthResponseObject, error)
	// Delete an application
	// (DELETE /v1/vm)
	DeleteVM(ctx context.Context, request DeleteVMRequestObject) (DeleteVMResponseObject, error)
	// Get a virtual machine application
	// (GET /v1/vm)
	GetVM(ctx context.Context, request GetVMRequestObject) (GetVMResponseObject, error)
	// Create a VM
	// (POST /v1/vm)
	CreateVM(ctx context.Context, request CreateVMRequestObject) (CreateVMResponseObject, error)
	// Update an application
	// (PUT /v1/vm)
	ApplyVM(ctx context.Context, request ApplyVMRequestObject) (ApplyVMResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// ListHealth operation middleware
func (sh *strictHandler) ListHealth(w http.ResponseWriter, r *http.Request) {
	var request ListHealthRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListHealth(ctx, request.(ListHealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListHealth")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListHealthResponseObject); ok {
		if err := validResponse.VisitListHealthResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// DeleteVM operation middleware
func (sh *strictHandler) DeleteVM(w http.ResponseWriter, r *http.Request) {
	var request DeleteVMRequestObject

	var body DeleteVMJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteVM(ctx, request.(DeleteVMRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteVM")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(DeleteVMResponseObject); ok {
		if err := validResponse.VisitDeleteVMResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetVM operation middleware
func (sh *strictHandler) GetVM(w http.ResponseWriter, r *http.Request) {
	var request GetVMRequestObject

	var body GetVMJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetVM(ctx, request.(GetVMRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetVM")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetVMResponseObject); ok {
		if err := validResponse.VisitGetVMResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateVM operation middleware
func (sh *strictHandler) CreateVM(w http.ResponseWriter, r *http.Request) {
	var request CreateVMRequestObject

	var body CreateVMJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateVM(ctx, request.(CreateVMRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateVM")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateVMResponseObject); ok {
		if err := validResponse.VisitCreateVMResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ApplyVM operation middleware
func (sh *strictHandler) ApplyVM(w http.ResponseWriter, r *http.Request) {
	var request ApplyVMRequestObject

	var body ApplyVMJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ApplyVM(ctx, request.(ApplyVMRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ApplyVM")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ApplyVMResponseObject); ok {
		if err := validResponse.VisitApplyVMResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}